feat: using loading and snackbar on root
