+++ src/app/login/page.tsx
@@ -3,7 +3,7 @@ import Checkbox from '@mui/material/Checkbox';
 import { Logo } from 'components';
 import Link from 'next/link';
 import { project } from 'shared/enum';
-import { InputPassword, InputText, SnackbarAlerts } from 'stories/components';
+import { InputPassword, InputText } from 'stories/components';
 import { Button } from 'stories/components/Forms/Button';
 import { getErrorsFormik, getHelperTextFormik } from 'utils';
 
@@ -11,7 +11,7 @@ import * as S from './styles';
 import { useLogin } from './useLogin';
 
 export default function Login() {
-  const { formik, loginError, setLoginError } = useLogin();
+  const { formik } = useLogin();
 
   return (
     <S.Main>
@@ -53,14 +53,6 @@ export default function Login() {
             Entrar
           </Button>
         </S.Form>
-
-        <SnackbarAlerts
-          setOpen={() => setLoginError('')}
-          open={!!loginError}
-          message={loginError}
-          severity={'error'}
-          onClose={() => setLoginError('')}
-        />
       </S.Wrapper>
     </S.Main>
   );
