+++ src/app/reset-password/page.tsx
@@ -3,28 +3,18 @@
 import { Logo } from 'components';
 import { redirect } from 'next/navigation';
 import { project } from 'shared/enum';
-import { Button, InputPassword, SnackbarAlerts } from 'stories/components';
+import { Button, InputPassword } from 'stories/components';
 import { getErrorsFormik, getHelperTextFormik } from 'utils';
 
 import * as S from './styles';
 import { useResetPassword } from './useResetPassword';
 
 export default function ResetPassword() {
-  const { formik, email, notification, setNotification, successMessage } = useResetPassword();
+  const { formik, email } = useResetPassword();
 
   if (!email) {
     redirect('/login');
   }
-
-  const getSeverity = () => {
-    if (notification) {
-      return 'error';
-    } else if (successMessage) {
-      return 'success';
-    } else {
-      return 'error';
-    }
-  };
 
   return (
     <S.Main>
@@ -65,13 +55,6 @@ export default function ResetPassword() {
             Enviar
           </Button>
         </S.Form>
-        <SnackbarAlerts
-          setOpen={() => setNotification('')}
-          open={!!notification || !!successMessage}
-          message={notification || successMessage}
-          severity={getSeverity()}
-          onClose={() => setNotification('')}
-        />
       </S.Wrapper>
     </S.Main>
   );
